feat: 帖子列表；帖子详情

+++ mk-community/src/main/java/com/github/xujiaji/mk/community/entity/MkCommunityArticle.java
@@ -63,7 +63,6 @@ public class MkCommunityArticle extends BaseEntity {
     /**
      * 收藏数量
      */
-    private Integer collectNum;
-
+    private Long collectNum;
 
 }
